+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.SoundSenseV1">
-        <activity android:name=".LoginActivity2"></activity>
+        <activity android:name=".LoginActivity"></activity>
         <activity android:name=".IntroductoryActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -22,8 +22,6 @@
             </intent-filter>
         </activity>
         <activity android:name=".PasswordActivity" />
-        <activity android:name=".RegistrationActivity" />
-        <activity android:name=".LoginActivity" />
         <activity
             android:name=".DataActivity"
             android:parentActivityName=".MainActivity" />
