--- app/src/main/java/com/example/soundsensev1/LoginActivity2.java
@@ -1,52 +1,0 @@
-package com.example.soundsensev1;
-
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.viewpager.widget.ViewPager;
-
-import android.os.Bundle;
-
-import com.google.android.material.tabs.TabLayout;
-
-public class LoginActivity2 extends AppCompatActivity {
-
-    TabLayout tabLayout;
-    ViewPager viewPager;
-    float v = 0;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_login2);
-
-        tabLayout = findViewById(R.id.tab_layout);
-        viewPager = findViewById(R.id.view_pager);
-
-        tabLayout.addTab(tabLayout.newTab().setText("Login"));
-        tabLayout.addTab(tabLayout.newTab().setText("SignUp"));
-        tabLayout.setTabGravity(TabLayout.GRAVITY_FILL);
-
-
-        final LoginAdapter adapter = new LoginAdapter(getSupportFragmentManager(), this, tabLayout.getTabCount());
-        viewPager.setAdapter(adapter);
-
-        viewPager.addOnPageChangeListener(new TabLayout.TabLayoutOnPageChangeListener(tabLayout));
-
-        tabLayout.addOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
-            @Override
-            public void onTabSelected(TabLayout.Tab tab) {
-                viewPager.setCurrentItem(tab.getPosition());
-            }
-            @Override
-            public void onTabUnselected(TabLayout.Tab tab) {
-            }
-            @Override
-            public void onTabReselected(TabLayout.Tab tab) {
-            }
-        });
-
-        //tabLayout.setTranslationY(300);
-        //tabLayout.setAlpha(v);
-
-    }
-
-}
